fix read_raw_eeg_file
data should not be hardcoded or stored in the code directory. These should be
generic functions that operate on paths defined by the user of the code

superfluous data is removed

global variable was removed. a good heuristic is that if you are using a
global variable you are doing something wrong

removed additional keyword arguments to the io function from mne. You only
specify keyword arguments if you want to. You don't have to specify a value
for them if you do not want.

do not use "pass". Not sure why it was included here