rename and improve tests file
test files should start with test to be picked up by pytest

You should run your tests using pytest

The creation of test data has been tidied into a function to
make troubleshooting it a little easier. Ultimately creating a
test fixture would be ideal for this sort of function.

Using Path objects from pathlib helps when working with lots of
directories and filenames. It adds the appropriate path seperators
"/" or "\" depending on your operating system. And you no longer have
to worry about trailing slashes. And it has lots of neat methods.

If you wish to create variables throughout a .py file the correct style
is to capitilize the names